+++ R/reexports.R
@@ -30,3 +30,7 @@ generics::fit
 #' @importFrom reticulate as_iterator
 #' @export
 reticulate::as_iterator
+
+#' @importFrom reticulate iter_next
+#' @export
+reticulate::iter_next