+++ tests/testthat/test-dataset-methods.R
@@ -173,6 +173,45 @@ test_succeeds("dataset_padded_batch returns a dataset", {
       padding_values = tf$constant(77L))
 
 
+test_succeeds("dataset_bucet_by_sequence_length", {
+  dataset <- list(c(0),
+                  c(1, 2, 3, 4),
+                  c(5, 6, 7),
+                  c(7, 8, 9, 10, 11),
+                  c(13, 14, 15, 16, 17, 18, 19, 20),
+                  c(21, 22)) %>%
+    lapply(as.array) %>% lapply(as_tensor, "int32") %>%
+    lapply(tensors_dataset) %>%
+    purrr::reduce(dataset_concatenate)
+
+  res <- dataset %>%
+    dataset_bucket_by_sequence_length(
+      element_length_func = function(elem) tf$shape(elem)[1],
+      bucket_boundaries = c(3, 5),
+      bucket_batch_sizes = c(2, 2, 2)
+    ) %>%
+    as_array_iterator() %>%
+    iterate(simplify = FALSE)
+
+  expected <-
+    list(structure(c(1L, 5L, 2L, 6L, 3L, 7L, 4L, 0L), .Dim = c(2L, 4L)),
+         structure(c(7L, 13L, 8L, 14L, 9L, 15L, 10L, 16L, 11L, 17L, 0L, 18L, 0L, 19L, 0L, 20L), .Dim = c(2L, 8L)),
+         structure(c(0L, 21L, 0L, 22L), .Dim = c(2L, 2L)))
+
+#' #      [,1] [,2] [,3] [,4]
+#' # [1,]    1    2    3    4
+#' # [2,]    5    6    7    0
+#' #      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
+#' # [1,]    7    8    9   10   11    0    0    0
+#' # [2,]   13   14   15   16   17   18   19   20
+#' #      [,1] [,2]
+#' # [1,]    0    0
+#' # [2,]   21   22
+
+  expect_identical(res, expected)
+})
+
+
 })
 
 test_succeeds("zip_datasets returns a dataset", {