+++ tests/testthat/test-dataset-methods.R
@@ -172,6 +172,43 @@ test_succeeds("dataset_padded_batch returns a dataset", {
       padded_shapes = tf$constant(3L, shape = shape(1L), dtype = tf$int32),
       padding_values = tf$constant(77L))
 
+})
+
+test_succeeds("dataset_padded_batch", {
+  A <- range_dataset(1, 5) %>%
+    dataset_map(function(x) tf$fill(list(x), x))
+
+  # Pad to the smallest per-batch size that fits all elements.
+  B <- A %>% dataset_padded_batch(2)
+  B %>% as_array_iterator() %>% iterate(simplify = FALSE) -> res
+  expected <- list(array(c(1, 2, 0, 2), c(2L, 2L)),
+                   array(c(3, 4, 3, 4, 3, 4, 0, 4), c(2L, 4L)))
+  expect_identical(res, expected)
+
+  # Pad to a fixed size.
+  C <- A %>% dataset_padded_batch(2, padded_shapes=5)
+  C %>% as_array_iterator() %>% iterate(simplify = FALSE) -> res
+  expected <- list(structure(c(1, 2, 0, 2, 0, 0, 0, 0, 0, 0), .Dim = c(2L, 5L)),
+                   structure(c(3, 4, 3, 4, 3, 4, 0, 4, 0, 0), .Dim = c(2L, 5L)))
+  expect_identical(res, expected)
+
+  # Pad with a custom value.
+  D <- A %>% dataset_padded_batch(2, padded_shapes=5, padding_values=-1)
+  D %>% as_array_iterator() %>% iterate(simplify = FALSE) -> res
+  expected <- list(structure(c(1, 2, -1, 2, -1, -1, -1, -1, -1, -1), .Dim = c(2L, 5L)),
+                   structure(c(3, 4, 3, 4, 3, 4, -1, 4, -1, -1), .Dim = c(2L, 5L)))
+  expect_identical(res, expected)
+
+  # Pad with a single value and multiple components.
+  E <- zip_datasets(A, A) %>%  dataset_padded_batch(2, padding_values = -1)
+  E %>% as_array_iterator() %>% iterate(simplify = FALSE) -> res
+  expected <- list(list(structure(c(1, 2, -1, 2), .Dim = c(2L, 2L)),
+                        structure(c(1,  2, -1, 2), .Dim = c(2L, 2L))),
+                   list(structure(c(3, 4, 3, 4, 3,  4, -1, 4), .Dim = c(2L, 4L)),
+                        structure(c(3, 4, 3, 4, 3, 4, -1,  4), .Dim = c(2L, 4L))))
+  expect_identical(res, expected)
+})
+
 
 test_succeeds("dataset_bucet_by_sequence_length", {
   dataset <- list(c(0),