+++ NEWS.md
@@ -1,3 +1,5 @@
+# tfdatasets (development version)
+
 # tfdatasets 2.6.0
 
 - New `dataset_options()` for setting and getting dataset options.